+++ src/com/spikes2212/genericsubsystems/drivetrains/HolonomicDrive.java
@@ -1,4 +1,4 @@
-package com.spikes2212.genericsubsystems.drivetrain;
+package com.spikes2212.genericsubsystems.drivetrains;
 
 public interface HolonomicDrive extends TankDrive {
 	public void holonomicDrive(double speedX, double speedY); // x-forward/backward
