+++ src/com/spikes2212/genericsubsystems/drivetrains/TankDrive.java
@@ -1,4 +1,4 @@
-package com.spikes2212.genericsubsystems.drivetrain;
+package com.spikes2212.genericsubsystems.drivetrains;
 
 public interface TankDrive {
 	public void tankDrive(double spLeft, double spRight);
